Move browseNewOutputPath to defaults client module
