+++ src/extension/src/client-modules/validator.ts
@@ -1,5 +1,3 @@
-import * as vscode from "vscode";
-import { CLI_SETTINGS } from "../constants/cli";
 import { EXTENSION_COMMANDS } from "../constants/commands";
 import fs = require("fs");
 import path = require("path");
@@ -8,33 +6,8 @@ import { MESSAGES } from "../constants/messages";
 
 export class Validator extends WizardServant {
   clientCommandMap: Map<EXTENSION_COMMANDS, (message: any) => Promise<IPayloadResponse>> = new Map([
-    [EXTENSION_COMMANDS.BROWSE_NEW_OUTPUT_PATH, Validator.browseNewOutputPath],
     [EXTENSION_COMMANDS.PROJECT_PATH_VALIDATION, Validator.handleProjectPathValidation],
   ]);
-
-  public static async browseNewOutputPath(message: any): Promise<IPayloadResponse> {
-    const openDialogConfig = { canSelectFiles: false, canSelectFolders: true, canSelectMany: false };
-    return vscode.window.showOpenDialog(openDialogConfig).then((response) => {
-      const outputPath = Validator.getOutputPath(response);
-      return {
-        payload: {
-          scope: message.payload.scope,
-          outputPath,
-        },
-      };
-    });
-  }
-
-  private static getOutputPath(path?: vscode.Uri[]): string | undefined {
-    if (path === undefined) {
-      return undefined;
-    }
-    let outputPath = path[0].path;
-    if (process.platform === CLI_SETTINGS.WINDOWS_PLATFORM_VERSION) {
-      outputPath = outputPath.substring(1, path[0].path.length);
-    }
-    return outputPath;
-  }
 
   public static async handleProjectPathValidation(message: any): Promise<IPayloadResponse> {
     const projectPath = message.projectPath;
